firmware: ti_sci: Correct the timeout type in ti_sci_do_xfer()
It should be 'unsigned long', not int.

+++ drivers/firmware/ti_sci.c
@@ -421,7 +421,7 @@ static inline int ti_sci_do_xfer(struct ti_sci_info *info,
 				 struct ti_sci_xfer *xfer)
 {
 	int ret;
-	int timeout;
+	unsigned long timeout;
 	struct device *dev = info->dev;
 
 	ret = mbox_send_message(info->chan_tx, &xfer->tx_message);
